fix: nickname 찾기 모달 분리
